# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Campus Entity Resolution API",
    "description": "API for campus security monitoring and entity tracking",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/anomalies/all": {
      "get": {
        "tags": ["anomaly-detection"],
        "summary": "Get All Historical Anomalies",
        "description": "Get all anomalies from the entire dataset with optional pagination",
        "operationId": "get_all_historical_anomalies_api_v1_anomalies_all_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "description": "Limit number of results (default: no limit)",
              "title": "Limit"
            },
            "description": "Limit number of results (default: no limit)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "description": "Offset for pagination",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/anomalies/date-range": {
      "get": {
        "tags": ["anomaly-detection"],
        "summary": "Get Anomalies By Date Range",
        "description": "Get anomalies within a specific date range",
        "operationId": "get_anomalies_by_date_range_api_v1_anomalies_date_range_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Start date in YYYY-MM-DD format",
              "title": "Start Date"
            },
            "description": "Start date in YYYY-MM-DD format"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "End date in YYYY-MM-DD format",
              "title": "End Date"
            },
            "description": "End date in YYYY-MM-DD format"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/anomalies/summary": {
      "get": {
        "tags": ["anomaly-detection"],
        "summary": "Get Anomaly Summary",
        "description": "Get summary of anomalies by type and severity",
        "operationId": "get_anomaly_summary_api_v1_anomalies_summary_get",
        "parameters": [
          {
            "name": "time_window_hours",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "description": "Time window in hours (if None, uses entire dataset)",
              "title": "Time Window Hours"
            },
            "description": "Time window in hours (if None, uses entire dataset)"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "description": "Start date in YYYY-MM-DD format",
              "title": "Start Date"
            },
            "description": "Start date in YYYY-MM-DD format"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "description": "End date in YYYY-MM-DD format",
              "title": "End Date"
            },
            "description": "End date in YYYY-MM-DD format"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/anomalies/by-location/{location}": {
      "get": {
        "tags": ["anomaly-detection"],
        "summary": "Get Anomalies By Location",
        "description": "Get anomalies for a specific location",
        "operationId": "get_anomalies_by_location_api_v1_anomalies_by_location__location__get",
        "parameters": [
          {
            "name": "location",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Location" }
          },
          {
            "name": "time_window_hours",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "description": "Time window in hours (if None, uses entire dataset)",
              "title": "Time Window Hours"
            },
            "description": "Time window in hours (if None, uses entire dataset)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/anomalies/by-severity/{severity}": {
      "get": {
        "tags": ["anomaly-detection"],
        "summary": "Get Anomalies By Severity",
        "description": "Get anomalies filtered by severity level",
        "operationId": "get_anomalies_by_severity_api_v1_anomalies_by_severity__severity__get",
        "parameters": [
          {
            "name": "severity",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Severity" }
          },
          {
            "name": "time_window_hours",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "description": "Time window in hours (if None, uses entire dataset)",
              "title": "Time Window Hours"
            },
            "description": "Time window in hours (if None, uses entire dataset)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/anomalies/by-entity/{entity_id}": {
      "get": {
        "tags": ["anomaly-detection"],
        "summary": "Get Anomalies By Entity",
        "description": "Get all anomalies for a specific entity (student/staff/faculty)",
        "operationId": "get_anomalies_by_entity_api_v1_anomalies_by_entity__entity_id__get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Entity Id" }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "description": "Start date in YYYY-MM-DD format",
              "title": "Start Date"
            },
            "description": "Start date in YYYY-MM-DD format"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "description": "End date in YYYY-MM-DD format",
              "title": "End Date"
            },
            "description": "End date in YYYY-MM-DD format"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}